+++ NEWS.md
@@ -18,6 +18,9 @@
 
 * `separate()` fills from right-to-left (not left-to-right!) when fill = "left"
   (#170, @dgrtwo).
+  
+* `separate_rows()` separates observations with multiple delimited values into
+  separate rows (#69, @aaronwolen).
 
 * `gather()` can now gather together list columns (#175).